% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_up.R
\name{eq_clean_data}
\alias{eq_clean_data}
\title{Takes raw NOAA data frame and returns a clean data frame.}
\usage{
eq_clean_data(data_raw)
}
\arguments{
\item{data_raw}{A data.frame or table specifying the raw data generated by
readr::read_delim}
}
\value{
A data.frame. The cleaned data.
}
\description{
The clean data frame should have the following: 1. A 'date' column created by
uniting the year, month, day and converting it to the Date class. 2. LATITUDE
and LONGITUDE columns converted to numeric class. 3. A cleaned LOCATION_NAME.
}
\examples{
\dontrun{
data_raw <- readr::read_delim("./signif.txt")
data_cleaned <- eq_clean_data(data_raw)
}
}
