% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaflet_map.R
\name{eq_create_label}
\alias{eq_create_label}
\title{Creates an HTML label for a data point.}
\usage{
eq_create_label(df)
}
\arguments{
\item{df}{A data.frame for which a list of labels will be generated. It must
have columns named 'LOCATION_NAME', 'EQ_PRIMARY', and 'TOTAL_DEATHS'.}
}
\value{
A list of characters, specifying the label for each data point.
}
\description{
\code{eq_create_label()} takes the dataset as an argument and creates an HTML
label that can be used as the annotation text in the leaflet map. This
function puts together a character string for each earthquake that will show
the cleaned location (cleaned LOCATION_NAME), the magnitude (EQ_PRIMARY),
and the total number of deaths (TOTAL_DEATHS), with boldface labels for
each ("Location", "Total deaths", and "Magnitude"). If an earthquake is
missing values for any of these, both the label and the value will be
skipped for that element of the tag.
}
\examples{
\dontrun{
data_raw <- readr::read_delim("./signif.txt")
dat <- eq_clean_data(data_raw) \%>\%
  filter(COUNTRY \%in\% c("MEXICO", "ITALY"))

label_1 <- eq_create_label(dat[301,])
# "<b>Location: </b>Guerrero<br><b>Magnitude: </b> 8.3<br><b>Total deaths: </b>8<br>"
}
}
