% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_up.R
\name{eq_location_clean}
\alias{eq_location_clean}
\title{Clean location names.}
\usage{
eq_location_clean(name_str)
}
\arguments{
\item{name_str}{A string specifying a LOCATION_NAME.}
}
\value{
A string. The cleaned location name.
}
\description{
For each string input, cleans it by stripping out the country name
(including the colon) and converts names to title case.
}
\examples{
\donttest{
loc_name <- "JORDAN:  BAB-A-DARAA,AL-KARAK"
loc_name_cleaned <- eq_location_clean(loc_name) # "Bab-A-Daraa,Al-Karak"
}
}
