% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaflet_map.R
\name{eq_map}
\alias{eq_map}
\title{Visualizes earthquakes.}
\usage{
eq_map(df, annot_col = "date")
}
\arguments{
\item{df}{A data.frame. The input earthquake data.}

\item{annot_col}{A string, specifying name of the column to be used for annotation
in the pop-up text. If "popup_text" is chosen, the annotation label will be a
complex label created by \code{eq_create_label()}. Default is "date".}
}
\value{
A leaflet. Contains an interactive map of earthquakes and annotations.
}
\description{
\code{eq_map()} takes an argument data containing the filtered data frame with
earthquakes to visualize. The function maps the epicenters (LATITUDE/LONGITUDE)
and annotates each point with in pop up window containing annotation data
stored in a column of the data frame. The user can choose which column is
used for the annotation in the pop-up with a function argument named
annot_col. Each earthquake should be shown with a circle, and the radius
of the circle should be proportional to the earthquake's magnitude (EQ_PRIMARY).
}
\examples{
\dontrun{
# create a leaflet map using 'date' as labels
readr::read_delim("./signif.txt") \%>\%
  eq_clean_data() \%>\%
  dplyr::filter(COUNTRY == "MEXICO" & lubridate::year(date) >= 2000) \%>\%
  eq_map(annot_col = "date")

# create a leaflet map using complex label
readr::read_delim("./signif.txt") \%>\%
  eq_clean_data() \%>\%
  dplyr::filter(COUNTRY == "MEXICO" & lubridate::year(date) >= 2000) \%>\%
  dplyr::mutate(popup_text = eq_create_label(.)) \%>\%
  eq_map(annot_col = "popup_text")
}
}
