% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_functions.R
\name{geom_timeline}
\alias{geom_timeline}
\title{The timeline geom plots a time line of earthquakes.}
\usage{
geom_timeline(
  mapping = NULL,
  data = NULL,
  stat = "timeline",
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  x_min = NULL,
  x_max = NULL,
  ...
)
}
\arguments{
\item{mapping}{A aesthetics class specifying mapping.}

\item{data}{A data.frame to compute the statistics from.}

\item{stat}{A string specifying a statistics class to use for plotting. Default is
'stat_timeline'.}

\item{position}{A string specifying position of plot.}

\item{na.rm}{A boolean value. Remove NA if 'TRUE'.}

\item{show.legend}{A boolean value. Show legend in plot if 'TRUE'.}

\item{inherit.aes}{A boolean value. Inherit aesthetics if 'TRUE'.}

\item{x_min}{A string specifying a date in a 'yyyy-mm-dd' format. Used as a
lowerbound to subset data. Default is NULL. Passed to stat_timeline if used.}

\item{x_max}{A string specifying a date in a 'yyyy-mm-dd' format. Used as a
upperbound to subset data. Default is NULL. Passed to stat_timeline if used.}

\item{...}{Other arguments passed on to layer()}
}
\value{
A ggplot2 layer.
}
\description{
\code{geom_timeline()} subsets the input data to a range specified by 'xmin' and/or
'xmax', if these parameters are supplied. Then it plots time line on which each
point is an earthquake. Optional aesthetics include color, size, and alpha
(for transparency). The xaesthetic is a date and an optional y aesthetic is
a factor indicating some stratification in which case multiple time lines
will be plotted for each level of the factor (e.g. country).
}
\examples{
\dontrun{
library(ggplot2)
# read data
data_raw <- readr::read_delim("./signif.txt")
dat <- eq_clean_data(data_raw) \%>\%
  filter(COUNTRY \%in\% c("MEXICO", "ITALY"))

# plot data on one timeline in uniform size and color
ggplot(dat, aes(x=date)) +
  geom_timeline()

# plot data in multiple timelines
ggplot(dat, aes(x=date, y=COUNTRY)) +
  geom_timeline()

# plot data in multiple timelines, specifying size by 'INTENSITY', and color by 'TOTAL_DEATHS'
ggplot(dat, aes(x=date, y=COUNTRY, colour=TOTAL_DEATHS, size=INTENSITY)) +
  geom_timeline()

# Use parameters 'alpha', 'x_min', and 'x_max'
ggplot(dat, aes(x=date, y=COUNTRY, colour=TOTAL_DEATHS, size=INTENSITY)) +
  geom_timeline(alpha=.1, x_min="1980-01-01", x_max="2017-01-01")
}
}
