% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_functions.R
\docType{data}
\name{StatTimeline}
\alias{StatTimeline}
\alias{GeomTimeline}
\alias{GeomTimelineLabel}
\title{stat_timeline constructor.}
\description{
stat_timeline constructor.

geom_timeline constructor.

geom_timeline_label constructor.
}
\details{
extends Stat

extends GeomPoint

extends GeomPoints
}
\keyword{datasets}
