% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_functions.R
\name{theme_timeline}
\alias{theme_timeline}
\title{The timeline theme for 'geom_timeline' or 'geom_timeline_label'.}
\usage{
theme_timeline()
}
\description{
\code{theme_timeline()} adds theme specifications to a ggplot2 layer created by
'geom_timeline' or 'geom_timeline_label'.
}
\examples{
\dontrun{
library(ggplot2)
data_raw <- readr::read_delim("./signif.txt")
dat <- eq_clean_data(data_raw) \%>\%
  filter(COUNTRY \%in\% c("MEXICO", "ITALY"))

# add theme
ggplot(dat, aes(x=date, y=COUNTYR, colour=TOTAL_DEATHS, size=INTENSITY)) +
  geom_timeline(alpha=.1, x_min="1980-01-01", x_max="2017-01-01") +
  theme_timeline() +
  xlab("DATE") +
  guides(size = guide_legend(title = "Richter scale value", order = 1),
    colour = guide_colorbar(title = "# deaths", order = 2))
}
}
